% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{clusterBar}
\alias{clusterBar}
\title{Create barplot of element concentration by cluster}
\usage{
clusterBar(dat_cluster_avg)
}
\description{
Create barplot of element concentration by cluster
}
