% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{clusterPotteryDat}
\alias{clusterPotteryDat}
\title{Cluster a data matrix using principle components}
\usage{
clusterPotteryDat(dat_vals, k, pc_keep = NULL, method = "gmm")
}
\description{
/code{clusterPotteryDat} takes in a data matrix, performs a principle
components transformation, and clusters using the first k components.
}
