% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{datClusterAvgFun}
\alias{datClusterAvgFun}
\title{Compute mean element concentration by cluster}
\usage{
datClusterAvgFun(dat_plot, elements)
}
\description{
Compute mean element concentration by cluster
}
