% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{processPotteryDat}
\alias{processPotteryDat}
\title{Process pottery data file}
\usage{
processPotteryDat(file_path, avg_obs = FALSE)
}
\description{
Takes in the output from read_csv(...) and processes according to
the conventionss
}
