% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transformDat}
\alias{transformDat}
\title{Extract and Transform Element Composition Values}
\usage{
transformDat(dat, elements, avg_readings = FALSE, standardize = FALSE,
  transformation = "none", offset = 1e-08)
}
\description{
\code{transformDat} extracts
}
